refactoring Tableau
