+++ src/Tableau/index.js
@@ -1,0 +1,11 @@
+/*global require*/
+/*global module*/
+require("./simplex.js");
+require("./cuttingStrategies.js");
+require("./dynamicModification.js");
+require("./log.js");
+require("./backup.js");
+require("./branchingStrategies.js");
+require("./integerProperties.js");
+
+module.exports = require("./Tableau.js");
