+++ src/Tableau/simplex.js
@@ -1,0 +1,322 @@
+/*global describe*/
+/*global require*/
+/*global module*/
+/*global it*/
+/*global console*/
+/*global process*/
+
+var Tableau = require("./Tableau.js");
+
+//-------------------------------------------------------------------
+// Function: solve
+// Detail: Main function, linear programming solver
+//-------------------------------------------------------------------
+Tableau.prototype.solve = function () {
+// this.log('INIT')
+    // Execute Phase 1 to obtain a Basic Feasible Solution (BFS)
+    this.phase1();
+
+    // Execute Phase 2
+    if (this.feasible === true) {
+        // Running simplex on Initial Basic Feasible Solution (BFS)
+        // N.B current solution is feasible
+        this.phase2();
+    }
+
+    return this;
+};
+
+//-------------------------------------------------------------------
+// Description: Convert a non standard form tableau
+//              to a standard form tableau by eliminating
+//              all negative values in the Right Hand Side (RHS)
+//              This results in a Basic Feasible Solution (BFS)
+//
+//-------------------------------------------------------------------
+Tableau.prototype.phase1 = function () {
+    var matrix = this.matrix;
+    var rhsColumn = this.rhsColumn;
+    var lastColumn = this.width - 1;
+    var lastRow = this.height - 1;
+
+    var unrestricted;
+    var iterations = 0;
+    while (true) {
+        // Selecting leaving variable (feasibility condition):
+        // Basic variable with most negative value
+        var leavingRowIndex = 0;
+        var rhsValue = -this.precision;
+        for (var r = 1; r <= lastRow; r++) {
+            unrestricted = this.unrestrictedVars[this.varIndexByRow[r]] === true;
+            if (unrestricted) {
+                continue;
+            }
+
+            var value = matrix[r][rhsColumn];
+            if (value < rhsValue) {
+                rhsValue = value;
+                leavingRowIndex = r;
+            }
+        }
+
+        // If nothing is strictly smaller than 0; we're done with phase 1.
+        if (leavingRowIndex === 0) {
+            // Feasible, champagne!
+            this.feasible = true;
+            return iterations;
+        }
+
+        // Selecting entering variable
+        var enteringColumn = 0;
+        var maxQuotient = -Infinity;
+        var costRow = matrix[0];
+        var leavingRow = matrix[leavingRowIndex];
+        for (var c = 1; c <= lastColumn; c++) {
+            var coefficient = leavingRow[c];
+            if (-this.precision < coefficient && coefficient < this.precision) {
+                continue;
+            }
+
+            unrestricted = this.unrestrictedVars[this.varIndexByCol[c]] === true;
+            if (unrestricted || coefficient < -this.precision) {
+                var quotient = -costRow[c] / coefficient;
+                if (maxQuotient < quotient) {
+                    maxQuotient = quotient;
+                    enteringColumn = c;
+                }
+            }
+        }
+
+        if (enteringColumn === 0) {
+            // Not feasible
+            this.feasible = false;
+            return iterations;
+        }
+
+        this.pivot(leavingRowIndex, enteringColumn);
+        iterations += 1;
+    }
+};
+
+//-------------------------------------------------------------------
+// Description: Apply simplex to obtain optimal soltuion
+//              used as phase2 of the simplex
+//
+//-------------------------------------------------------------------
+Tableau.prototype.phase2 = function () {
+    var matrix = this.matrix;
+    var rhsColumn = this.rhsColumn;
+    var lastColumn = this.width - 1;
+    var lastRow = this.height - 1;
+
+    var precision = this.precision;
+    var nOptionalObjectives = this.optionalObjectives.length;
+    var optionalCostsColumns = null;
+
+    var iterations = 0;
+    var reducedCost, unrestricted;
+    while (true) {
+        var costRow = matrix[this.costRowIndex];
+
+        // Selecting entering variable (optimality condition)
+        if (nOptionalObjectives > 0) {
+            optionalCostsColumns = [];
+        }
+
+        var enteringColumn = 0;
+        var enteringValue = this.precision;
+        var isReducedCostNegative = false;
+        for (var c = 1; c <= lastColumn; c++) {
+            reducedCost = costRow[c];
+            unrestricted = this.unrestrictedVars[this.varIndexByCol[c]] === true;
+
+            if (nOptionalObjectives > 0 && -this.precision < reducedCost && reducedCost < this.precision) {
+                optionalCostsColumns.push(c);
+                continue;
+            }
+
+            if (unrestricted && reducedCost < 0) {
+                if (-reducedCost > enteringValue) {
+                    enteringValue = -reducedCost;
+                    enteringColumn = c;
+                    isReducedCostNegative = true;
+                }
+                continue;
+            }
+
+            if (reducedCost > enteringValue) {
+                enteringValue = reducedCost;
+                enteringColumn = c;
+                isReducedCostNegative = false;
+            }
+        }
+
+        if (nOptionalObjectives > 0) {
+            // There exist optional improvable objectives
+            var o = 0;
+            while (enteringColumn === 0 && optionalCostsColumns.length > 0 && o < nOptionalObjectives) {
+                var optionalCostsColumns2 = [];
+                var reducedCosts = this.optionalObjectives[o].reducedCosts;
+                for (var i = 0; i <= optionalCostsColumns.length; i++) {
+                    c = optionalCostsColumns[i];
+                    reducedCost = reducedCosts[c];
+                    unrestricted = this.unrestrictedVars[this.varIndexByCol[c]] === true;
+
+                    if (-this.precision < reducedCost && reducedCost < this.precision) {
+                        optionalCostsColumns2.push(c);
+                        continue;
+                    }
+
+                    if (unrestricted && reducedCost < 0) {
+                        if (-reducedCost > enteringValue) {
+                            enteringValue = -reducedCost;
+                            enteringColumn = c;
+                            isReducedCostNegative = true;
+                        }
+                        continue;
+                    }
+
+                    if (reducedCost > enteringValue) {
+                        enteringValue = reducedCost;
+                        enteringColumn = c;
+                        isReducedCostNegative = false;
+                    }
+                }
+                optionalCostsColumns = optionalCostsColumns2;
+                o += 1;
+            }
+        }
+
+        // If no entering column could be found we're done with phase 2.
+        if (enteringColumn === 0) {
+            this.setEvaluation();
+            return;
+        }
+
+        // Selecting leaving variable
+        var leavingRow = 0;
+        var minQuotient = Infinity;
+
+        for (var r = 1; r <= lastRow; r++) {
+            var row = matrix[r];
+            var rhsValue = row[rhsColumn];
+            var colValue = row[enteringColumn];
+
+            if (-precision < colValue && colValue < precision) {
+                continue;
+            }
+
+            if (colValue > 0 && precision > rhsValue && rhsValue > -precision) {
+                minQuotient = 0;
+                leavingRow = r;
+                break;
+            }
+
+            var quotient = isReducedCostNegative ? -rhsValue / colValue : rhsValue / colValue;
+            if (quotient > 0 && minQuotient > quotient) {
+                minQuotient = quotient;
+                leavingRow = r;
+            }
+        }
+
+        if (minQuotient === Infinity) {
+            // TODO: solution is not bounded
+            // optimal value is -Infinity
+            this.evaluation = -Infinity;
+            return;
+        }
+
+        this.pivot(leavingRow, enteringColumn, true);
+        iterations += 1;
+    }
+};
+
+// Array holding the column indexes for which the value is not null
+// on the pivot row
+// Shared by all tableaux for smaller overhead and lower memory usage
+var nonZeroColumns = [];
+//-------------------------------------------------------------------
+// Description: Execute pivot operations over a 2d array,
+//          on a given row, and column
+//
+//-------------------------------------------------------------------
+Tableau.prototype.pivot = function (pivotRowIndex, pivotColumnIndex) {
+    var matrix = this.matrix;
+
+    var quotient = matrix[pivotRowIndex][pivotColumnIndex];
+
+    var lastRow = this.height - 1;
+    var lastColumn = this.width - 1;
+
+    var leavingBasicIndex = this.varIndexByRow[pivotRowIndex];
+    var enteringBasicIndex = this.varIndexByCol[pivotColumnIndex];
+
+    this.varIndexByRow[pivotRowIndex] = enteringBasicIndex;
+    this.varIndexByCol[pivotColumnIndex] = leavingBasicIndex;
+
+    this.rowByVarIndex[enteringBasicIndex] = pivotRowIndex;
+    this.rowByVarIndex[leavingBasicIndex] = -1;
+
+    this.colByVarIndex[enteringBasicIndex] = -1;
+    this.colByVarIndex[leavingBasicIndex] = pivotColumnIndex;
+
+    // Divide everything in the target row by the element @
+    // the target column
+    var pivotRow = matrix[pivotRowIndex];
+    var nNonZeroColumns = 0;
+    for (var c = 0; c <= lastColumn; c++) {
+        if (pivotRow[c] !== 0) {
+            pivotRow[c] /= quotient;
+            nonZeroColumns[nNonZeroColumns] = c;
+            nNonZeroColumns += 1;
+        }
+    }
+    pivotRow[pivotColumnIndex] = 1 / quotient;
+
+    // for every row EXCEPT the pivot row,
+    // set the value in the pivot column = 0 by
+    // multiplying the value of all elements in the objective
+    // row by ... yuck... just look below; better explanation later
+    var coefficient, i, v0;
+    var precision = this.precision;
+    for (var r = 0; r <= lastRow; r++) {
+        var row = matrix[r];
+        if (r !== pivotRowIndex) {
+            coefficient = row[pivotColumnIndex];
+            // No point Burning Cycles if
+            // Zero to the thing
+            if (coefficient !== 0) {
+                for (i = 0; i < nNonZeroColumns; i++) {
+                    c = nonZeroColumns[i];
+                    // No point in doing math if you're just adding
+                    // Zero to the thing
+                    v0 = pivotRow[c];
+                    if (v0 !== 0) {
+                        row[c] = row[c] - coefficient * v0;
+                    }
+                }
+
+                row[pivotColumnIndex] = -coefficient / quotient;
+            }
+        }
+    }
+
+    var nOptionalObjectives = this.optionalObjectives.length;
+    if (nOptionalObjectives > 0) {
+        for (var o = 0; o < nOptionalObjectives; o += 1) {
+            var reducedCosts = this.optionalObjectives[o].reducedCosts;
+            coefficient = reducedCosts[pivotColumnIndex];
+            if (coefficient !== 0) {
+                for (i = 0; i < nNonZeroColumns; i++) {
+                    c = nonZeroColumns[i];
+                    v0 = pivotRow[c];
+                    if (v0 !== 0) {
+                        reducedCosts[c] = reducedCosts[c] - coefficient * v0;
+                    }
+                }
+
+                reducedCosts[pivotColumnIndex] = -coefficient / quotient;
+            }
+        }
+    }
+};
