+++ test/problems/Degenerate Min.json
@@ -1,0 +1,47 @@
+{
+        "name": "Degenerate Min",
+        "optimize": "objective",
+        "background": "http://arxiv.org/pdf/1304.2107.pdf",
+        "opType": "min",
+        "constraints": {
+            "c1": {
+                "max": 4
+            },
+            "c2": {
+                "min": 6
+            },
+            "c3": {
+                "min": 8
+            },
+            "c4": {
+                "min": 18
+            },
+            "c5": {
+                "min": 32
+            }
+        },
+        "variables": {
+            "x1": {
+                "c1": 1,
+                "c2": 0,
+                "c3": 1,
+                "c4": 3,
+                "c5": 5,
+                "objective": 3
+            },
+            "x2": {
+                "c1": 0,
+                "c2": 1,
+                "c3": 1,
+                "c4": 2,
+                "c5": 4,
+                "objective": 5
+            }
+        },
+        "expects": {
+            "feasible": true,
+            "x1": 2,
+            "x2": 6,
+            "result": 36
+        }
+    }
