+++ test/problems/Fertilizer.json
@@ -1,0 +1,54 @@
+{
+        "name": "Fertilizer",
+        "constraints" : {
+            "W": { "equal": 40, "weight": 1, "priority": 1 },
+            "N": { "equal": 50, "weight": 1, "priority": 0 },
+            "P": { "equal": 20, "weight": 2, "priority": 0 },
+            "K": { "equal": 30, "weight": 1, "priority": 0 }
+        },
+        "variables" : {
+            "Fertilizer A": { "N" : 36.00, "P" : 0.00,  "K" : 0.00,  "W" : 40 },
+            "Fertilizer B": { "N" : 15.00, "P" : 15.00, "K" : 15.00, "W" : 40 },
+            "Fertilizer C": { "N" : 5.00,  "P" : 25.00, "K" : 10.00, "W" : 40 }
+        },
+        "expects": {
+            "feasible": true,
+            "r1": 46.66666667,
+            "r4": 10,
+            "Fertilizer A": 0.83333333,
+            "Fertilizer B": 1.33333333,
+            "result": 10
+        }
+    }, {
+        "name": "Coffee Problem",
+        "optimize": "profit",
+        "opType": "max",
+        "constraints": {
+            "Costa Rican": {
+                "max": 200
+            },
+            "Etheopian": {
+                "max": 330
+            }
+        },
+        "variables": {
+            "Yusip": {
+                "Costa Rican": 0.5,
+                "Etheopian": 0.5,
+                "profit": 3.5,
+                "Yusip": 1
+            },
+            "Exotic": {
+                "Costa Rican": 0.25,
+                "Etheopian": 0.75,
+                "profit": 4,
+                "Exotic": 1
+            }
+        },
+        "expects": {
+            "feasible": true,
+            "Yusip": 270,
+            "Exotic": 260,
+            "result": 1985
+        }
+    }
