+++ test/problems/Generic Business Problem.json
@@ -1,0 +1,33 @@
+{
+        "name": "Generic Business Problem",
+        "optimize": "profit",
+        "opType": "max",
+        "constraints": {
+            "size": {
+                "max": 2400
+            },
+            "units": {
+                "min": 50
+            }
+        },
+        "variables": {
+            "large": {
+                "size": 60,
+                "units": 1,
+                "profit": 250,
+                "large": 1
+            },
+            "small": {
+                "size": 40,
+                "units": 1,
+                "profit": 150,
+                "small": 1
+            }
+        },
+        "expects": {
+            "feasible": true,
+            "large": 20,
+            "small": 30,
+            "result": 9500
+        }
+    }
