+++ test/problems/Integer Sports Complex Problem.json
@@ -1,0 +1,65 @@
+{
+        "name": "Integer Sports Complex Problem",
+        "optimize": "usage",
+        "opType": "max",
+        "constraints": {
+            "cost": {
+                "max": 120000
+            },
+            "land": {
+                "max": 12
+            },
+            "pool": {
+                "max": 1
+            },
+            "tennis": {
+                "max": 1
+            },
+            "field": {
+                "max": 1
+            },
+            "gym": {
+                "max": 1
+            }
+        },
+        "variables": {
+            "pool": {
+                "land": 4,
+                "cost": 35000,
+                "usage": 300,
+                "pool": 1
+            },
+            "tennis": {
+                "land": 2,
+                "cost": 10000,
+                "usage": 90,
+                "tennis": 1
+            },
+            "field": {
+                "land": 7,
+                "cost": 25000,
+                "usage": 400,
+                "field": 1
+            },
+            "gym": {
+                "land": 3,
+                "cost": 90000,
+                "usage": 150,
+                "gym": 1
+            }
+        },
+        "ints": {
+            "pool": 1,
+            "tennis": 1,
+            "field": 1,
+            "gym": 1
+        },
+        "expects": {
+            "feasible": true,
+            "pool": 1,
+            "tennis": 0,
+            "field": 1,
+            "gym": 0,
+            "result": 700
+        }
+    }
